+++ HISTORY.rst
@@ -7,6 +7,9 @@ Unreleased
 ++++++++++
 
 **Data Migration**
+* ALCN-637: Import pharmacist id in customer.
+* ALCN-638: Customer active flag.
+* ALCN-639: Import product tracking configuration.
 
 **Features and Improvements**